functions sayfasının ayarlanması
